refactor: uses kebab-case for directories
